Moving the predict() method implementation to the base DataRun class - it's children are now expected to implement the tel_pos_to_altaz() method to make predict() work.
